pyproject.toml
[project]
name = "idi"
version = "0.1.0"
description = "Intelligent Daemon Interface - ZK proof integration for Tau Net"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "zstandard>=0.22.0",
    "cryptography>=41.0.0",
    "aiohttp>=3.9.2",
    "PyYAML>=6.0.1",
    "py-ecc>=7.0.0",
]

[project.scripts]
ian-demo = "idi.ian.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "mutmut>=2.4.0",
    "pandas>=2.0.0",
    "hypothesis>=6.0.0",
    "pip-audit>=2.9.0",
    "z3-solver>=4.12.5.0",
]
gui = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
]
security = [
    "keyring>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-packages = true

[tool.hatch.build.targets.wheel.force-include]
"idi/ian/network/openapi.yaml" = "idi/ian/network/openapi.yaml"
"idi/ian/deploy/config" = "idi/ian/deploy/config"

[tool.pytest.ini_options]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
norecursedirs = ["mutants"]
markers = [
    "asyncio: mark a test as requiring an asyncio event loop (pytest-asyncio)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]  # Line length handled by formatter

[tool.coverage.run]
source = ["idi"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.mutmut]
paths_to_mutate = [
    "idi/zk/proof_manager.py",
    "idi/zk/bundle_verify.py",
    "idi/zk/wire.py",
    "idi/zk/verification.py",
]
# Note: mutmut runs the runner from the 'mutants' directory.
# Use a relative path back to the real tests.
runner = "../.venv/bin/python -m pytest ../idi/zk/tests -q"
